pyproject.toml
[project]
name = "260326-tti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "matplotlib>=3.9.4",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.3",
]
